requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt --resolver=backtracking pyproject.toml
#
absl-py==1.4.0
    # via
    #   tensorboard
    #   tensorflow
aiohttp==3.8.4
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
appnope==0.1.3
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
asttokens==2.2.1
    # via
    #   Chairdelure (pyproject.toml)
    #   stack-data
astunparse==1.6.3
    # via tensorflow
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via
    #   aiohttp
    #   pytest
backcall==0.2.0
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
black==23.1.0
    # via Chairdelure (pyproject.toml)
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via
    #   Chairdelure (pyproject.toml)
    #   pipenv
    #   requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via black
contourpy==1.0.7
    # via
    #   Chairdelure (pyproject.toml)
    #   matplotlib
coverage[toml]==7.2.2
    # via pytest-cov
cycler==0.11.0
    # via
    #   Chairdelure (pyproject.toml)
    #   matplotlib
decorator==5.1.1
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
discord-py==2.2.2
    # via Chairdelure (pyproject.toml)
distlib==0.3.6
    # via
    #   Chairdelure (pyproject.toml)
    #   virtualenv
executing==1.2.0
    # via
    #   Chairdelure (pyproject.toml)
    #   stack-data
filelock==3.9.0
    # via
    #   Chairdelure (pyproject.toml)
    #   virtualenv
flatbuffers==23.3.3
    # via tensorflow
fonttools==4.39.0
    # via
    #   Chairdelure (pyproject.toml)
    #   matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
gast==0.4.0
    # via tensorflow
google-auth==2.16.2
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.51.3
    # via
    #   tensorboard
    #   tensorflow
h5py==3.8.0
    # via tensorflow
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
ipython==8.11.0
    # via
    #   Chairdelure (pyproject.toml)
    #   nbqa
jax==0.4.6
    # via tensorflow
jedi==0.18.2
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
joblib==1.2.0
    # via
    #   Chairdelure (pyproject.toml)
    #   scikit-learn
keras==2.12.0rc1
    # via tensorflow
kiwisolver==1.4.4
    # via
    #   Chairdelure (pyproject.toml)
    #   matplotlib
libclang==15.0.6.1
    # via tensorflow
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.2
    # via werkzeug
matplotlib==3.7.1
    # via Chairdelure (pyproject.toml)
matplotlib-inline==0.1.6
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
nbqa==1.6.3
    # via Chairdelure (pyproject.toml)
numpy==1.23.5
    # via
    #   Chairdelure (pyproject.toml)
    #   contourpy
    #   h5py
    #   jax
    #   matplotlib
    #   opt-einsum
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via
    #   jax
    #   tensorflow
packaging==23.0
    # via
    #   Chairdelure (pyproject.toml)
    #   black
    #   matplotlib
    #   pytest
    #   tensorflow
pandas==1.5.3
    # via Chairdelure (pyproject.toml)
parso==0.8.3
    # via
    #   Chairdelure (pyproject.toml)
    #   jedi
pathspec==0.11.1
    # via black
pexpect==4.8.0
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
pickleshare==0.7.5
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
pillow==9.4.0
    # via
    #   Chairdelure (pyproject.toml)
    #   matplotlib
pipenv==2023.2.18
    # via Chairdelure (pyproject.toml)
platformdirs==3.1.1
    # via
    #   Chairdelure (pyproject.toml)
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
protobuf==4.22.1
    # via
    #   tensorboard
    #   tensorflow
ptyprocess==0.7.0
    # via
    #   Chairdelure (pyproject.toml)
    #   pexpect
pure-eval==0.2.2
    # via
    #   Chairdelure (pyproject.toml)
    #   stack-data
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygments==2.14.0
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
pyparsing==3.0.9
    # via
    #   Chairdelure (pyproject.toml)
    #   matplotlib
pytest==7.2.2
    # via
    #   Chairdelure (pyproject.toml)
    #   pytest-cov
pytest-cov==4.0.0
    # via Chairdelure (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   Chairdelure (pyproject.toml)
    #   matplotlib
    #   pandas
python-dotenv==0.21.1
    # via Chairdelure (pyproject.toml)
pytz==2022.7.1
    # via
    #   Chairdelure (pyproject.toml)
    #   pandas
requests==2.28.2
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
ruff==0.0.257
    # via Chairdelure (pyproject.toml)
scikit-learn==1.2.1
    # via Chairdelure (pyproject.toml)
scipy==1.10.1
    # via
    #   Chairdelure (pyproject.toml)
    #   jax
    #   scikit-learn
six==1.16.0
    # via
    #   Chairdelure (pyproject.toml)
    #   asttokens
    #   astunparse
    #   google-auth
    #   google-pasta
    #   python-dateutil
    #   tensorflow
stack-data==0.6.2
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
tensorboard==2.12.0
    # via tensorflow
tensorboard-data-server==0.7.0
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.12.0rc1
    # via Chairdelure (pyproject.toml)
tensorflow-estimator==2.12.0rc0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.31.0
    # via tensorflow
termcolor==2.2.0
    # via tensorflow
threadpoolctl==3.1.0
    # via
    #   Chairdelure (pyproject.toml)
    #   scikit-learn
tokenize-rt==5.0.0
    # via
    #   Chairdelure (pyproject.toml)
    #   nbqa
tomli==2.0.1
    # via
    #   Chairdelure (pyproject.toml)
    #   nbqa
traitlets==5.9.0
    # via
    #   Chairdelure (pyproject.toml)
    #   ipython
    #   matplotlib-inline
typing-extensions==4.5.0
    # via tensorflow
urllib3==1.26.15
    # via requests
virtualenv==20.20.0
    # via
    #   Chairdelure (pyproject.toml)
    #   pipenv
virtualenv-clone==0.5.7
    # via
    #   Chairdelure (pyproject.toml)
    #   pipenv
wcwidth==0.2.6
    # via
    #   Chairdelure (pyproject.toml)
    #   prompt-toolkit
werkzeug==2.2.3
    # via tensorboard
wheel==0.40.0
    # via
    #   astunparse
    #   tensorboard
wrapt==1.14.1
    # via tensorflow
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
